dev_requirements.txt
APScheduler==3.7.0
bcrypt==3.2.0
fastapi==0.68.2
Jinja2==2.11.3
pigpio==1.78
python-mpv==0.5.2
python-multipart==0.0.5
requests==2.25.1
tortoise-orm==0.17.2
uvicorn~=0.13.4
passlib~=1.7.4
python-jose
pytz
smbus2~=0.4.1
pigpio~=1.78
pimoroni-bme280~=0.1.1
Flask~=1.1.2
aioredis<2.0.0
